really fix SLES 10/11 mkinitrd

+++ jobslave/bootloader/grub_installer.py
@@ -366,20 +366,24 @@ class GrubInstaller(bootloader.BootloaderInstaller):
         kpaths = ['vmlinuz-' + x for x in kernels]
         ipaths = ['initrd-' + x for x in kernels]
 
-        # More SLES 11 magic: make a temporary device node
-        # for the root fs device, and remove it after mkinitrd runs.
-        tmpRootDev = os.path.join(self.image_root, 'tmp', 'root')
-        os.mknod(tmpRootDev, 0600 | stat.S_IFBLK, 
-                 os.stat(self.image_root).st_dev)
-
-        logCall(['/usr/sbin/chroot', self.image_root,
+        mkinitrdCmd = ['/usr/sbin/chroot', self.image_root,
             '/sbin/mkinitrd',
             '-k', ' '.join(kpaths),
             '-i', ' '.join(ipaths),
-            '-d', '/tmp/root',
-            ])
+            ]
+       
+        # More SLES 11 magic: make a temporary device node
+        # for the root fs device, and remove it after mkinitrd runs.
+        if is_SUSE(self.image_root, version=11):
+            tmpRootDev = os.path.join(self.image_root, 'dev', 'root')
+            os.mknod(tmpRootDev, 0600 | stat.S_IFBLK, 
+                     os.stat(self.image_root).st_dev)
+            mkinitrdCmd.extend([ '-d', '/dev/root' ])
 
-        os.unlink(tmpRootDev)
+        logCall(mkinitrdCmd)
+
+        if is_SUSE(self.image_root, version=11):
+            os.unlink(tmpRootDev)
 
         # Build grub config
         log.info("Adding kernel entries")
